// @cloud-materials/common/components/CForm/interface.ts
type ExcludeUndefined<T> = Exclude<T, undefined>;

/**
 * 默认的`CForm`内置组件map
 */
// ❓为什么要显式定义类型？
// 显式定义类型可以让引用了这个变量的地方生成d.ts时不用内联它们而是通过import("../CForm/interface").DefaultBuiltInComponentMapType这种形式
// 否则如bit这样的库在生成类型时会自动生成内联的类型，造成不同版本的物料出现类型错误
export type DefaultBuiltInComponentMapType = {
  Input: typeof Input;
  TextArea: ExcludeUndefined<typeof Input.TextArea>;
  InputSearch: ExcludeUndefined<typeof Input.Search>;
  Password: ExcludeUndefined<typeof Input.Password>;
  FormItem: typeof FormItem;
  InputNumber: typeof InputNumber;
  Section: typeof CFormSection;
  DynamicFieldDecorator: typeof DynamicFieldDecorator;
  CRadio: typeof CRadio;
  CCheckbox: typeof CCheckbox;
  Switch: typeof Switch;
  Space: typeof Space;
  ConfigPreview: typeof ConfigPreview;
  Select: typeof Select;
  Text: typeof Text;
  Cascader: typeof Cascader;
  Checkbox: typeof Checkbox;
  CheckboxGroup: ExcludeUndefined<typeof Checkbox.Group>;
  DatePicker: typeof DatePicker;
  WeekPicker: ExcludeUndefined<typeof DatePicker.WeekPicker>;
  MonthPicker: ExcludeUndefined<typeof DatePicker.MonthPicker>;
  YearPicker: ExcludeUndefined<typeof DatePicker.YearPicker>;
  QuarterPicker: ExcludeUndefined<typeof DatePicker.QuarterPicker>;
  RangePicker: ExcludeUndefined<typeof DatePicker.RangePicker>;
  TreeSelect: typeof TreeSelect;
  FeeCalculator: typeof FeeCalculator;
  Agreement: typeof CAgreement;
  CAgreement: typeof CAgreement;
  CAsyncSelect: typeof CAsyncSelect;
  TimePicker: typeof TimePicker;
  TimeRangePicker: ExcludeUndefined<typeof TimePicker.RangePicker>;
  Upload: typeof Upload;
  ArrayTable: typeof JSXTable;
  ArrayTableColumn: ExcludeUndefined<typeof ArrayTable.Column>;
  ArraySortHandle: ExcludeUndefined<typeof ArrayBase.SortHandle>;
  ArrayTableAddition: typeof ArrayTableAddition;
  ArrayIndex: ExcludeUndefined<typeof ArrayBase.Index>;
  ArrayMoveDown: ExcludeUndefined<typeof ArrayBase.MoveDown>;
  ArrayMoveUp: ExcludeUndefined<typeof ArrayBase.MoveUp>;
  ArrayTableRemove: typeof CArrayRemove;
  ArrayItems: typeof JSXItems;
  ArrayItemsRemove: typeof CArrayRemove;
  ArrayItemsAddition: typeof ArrayItemsAddition;
  Group: typeof CGroup;
  CTableSelect: typeof CTableSelect;
  PreSelectedTable: typeof PreSelectedTable;
  CDrawerSelect: typeof CFormDrawerSelect;
  CTableTransfer: typeof CFormTableTransfer;
  CFormAsyncSelect: typeof CFormAsyncSelectReactive;
};
// ...
export type VoidFieldType<
  T extends ObjectType = any,
  D extends ObjectType = any,
  ComponentsMap extends PureComponentsMap = DefaultBuiltInComponentMapType,
> = {
  type: 'VoidField';
} & CVoidFieldProps<T, D, ComponentsMap>;

export type BasicFieldType<
  T extends ObjectType = any,
  D extends ObjectType = any,
  ComponentsMap extends PureComponentsMap = DefaultBuiltInComponentMapType,
> = { type?: 'Field' | 'ArrayField' | 'ObjectField' } & CFieldProps<T, D, ComponentsMap>;

export type CFieldConfig<
  T extends ObjectType = any,
  D extends ObjectType = any,
  ComponentsMap extends PureComponentsMap = DefaultBuiltInComponentMapType,
> = VoidFieldType<T, D, ComponentsMap> | BasicFieldType<T, D, ComponentsMap>;
// ...
export interface CFieldProps<
  T extends ObjectType = any,
  D extends ObjectType = any,
  ComponentsMap extends PureComponentsMap = DefaultBuiltInComponentMapType,
> extends Omit<IFieldProps, BaseOmitKey | 'dataSource'>,
    CBaseFieldProps<T, D, ComponentsMap> {
  /**
   * @zh 字段数据源配置（注意，仅数据型字段支持dataSource属性配置）
   */
  dataSource?: CallableType<FieldDataSource, T, D>;
  /**
   * @zh 字段校验配置（注意，仅数据型字段支持rules属性配置）
   */
  rules?: RulesProps<T, D> | RulesProps<T, D>[];
}

export interface CVoidFieldProps<
  T extends ObjectType = any,
  D extends ObjectType = any,
  ComponentsMap extends PureComponentsMap = DefaultBuiltInComponentMapType,
> extends Omit<IVoidFieldProps, BaseOmitKey>,
    CBaseFieldProps<T, D, ComponentsMap> {}
// ...
export type ObjectType = Record<string, any>;

export type JSXComponent = keyof JSX.IntrinsicElements | React.JSXElementConstructor<any>;

export type ComponentsMap = Record<string, JSXComponent>;

export type PureComponentsMap = Record<string, React.JSXElementConstructor<any>>;
// ...
export type BaseOmitKey =
  | 'hidden'
  | 'visible'
  | 'display'
  | 'pattern'
  | 'editable'
  | 'disabled'
  | 'readOnly'
  | 'readPretty'
  | 'data'
  | 'decorator'
  | 'decoratorProps'
  | 'component'
  | 'componentProps'
  | 'title'
  | 'description'
  | 'validator';